Update data-from-other-node.md
Added the option to copy over Electrs index. This solves the problem of reindexing errors using the builtin Electrs server which have been experienced when using blockchain data from another device. This code will be different if Electrs was not the previous indexer or the desired indexer in the destination node.